websockets/pyproject.toml
[tool.poetry]
name = "project-stargate-websocket"
version = "0.1.0"
description = "FastAPI WebSocket for Project Stargate"
authors = ["Kurtis Massey <55586356+kurtismassey@users.noreply.github.com>"]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.109.1"
uvicorn = "^0.24.0"
websockets = "^12.0"
google-generativeai = "^0.7.2"
python-dotenv = "^1.0.1"
langchain-google-genai = "^1.0.8"
pillow = "^10.4.0"
langchain = "^0.2.12"
langchain-google-firestore = "^0.3.0"
google-cloud-aiplatform = "^1.61.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"